Delete unused keccak circuit; rename keccak2 to keccak (#191)
Confusion abounds
Two circuits for the same hash
Ya ain't gon' need it

We currently have two versions of the keccak: the "narrow" version, keccak, and the "wide" version, keccak2. The latter is the one we use in production, while the former is currently unused. Having these two separate implementations side-by-side has caused confusion for a number of external parties, so it seems best to clean this up. The source code for "narrow" keccak will be preserved in the git history, so if we decide we want it down the road we still have it.